requirements.txt
pandas
httpx
python-dotenv
langchain_groq
pydantic
fastapi
uvicorn
aiofiles
langchain_google_genai